requirements.txt
Flask
Werkzeug
PyMySQL
numpy
matplotlib
APScheduler
python-alipay-sdk
qrcode